fix: added a test for the prompt and refactored the dialog test
