+++ web/src/components/__tests__/Dialog.test.jsx
@@ -16,23 +16,8 @@ describe('Dialog', () => {
   });
 
   test('renders to a portal', async () => {
-    render(<Dialog title="Tacos" text="This is the dialog" />);
-    expect(screen.getByText('Tacos')).toBeInTheDocument();
+    render(<Dialog>Sample</Dialog>);
+    expect(screen.getByText('Sample')).toBeInTheDocument();
     expect(screen.getByRole('modal').closest('#dialogs')).not.toBeNull();
-  });
-
-  test('renders action buttons', async () => {
-    const handleClick = jest.fn();
-    render(
-      <Dialog
-        actions={[
-          { color: 'red', text: 'Delete' },
-          { text: 'Okay', onClick: handleClick },
-        ]}
-        title="Tacos"
-      />
-    );
-    fireEvent.click(screen.getByRole('button', { name: 'Okay' }));
-    expect(handleClick).toHaveBeenCalled();
   });
 });
